Assignment :
Part A :

2. Only the Non-owner of the image can delete funcationlaity has been implem

+++ ImageHoster/src/main/java/ImageHoster/controller/ImageController.java
@@ -152,9 +152,25 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session)
+    {
+        User loggedInUser = (User) session.getAttribute("loggeduser");
+        Image image = imageService.getImage(imageId);
+        User ownerOfTheImage = image.getUser();
+        if(loggedInUser.getId() == ownerOfTheImage.getId())
+        {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else
+        {
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("deleteError", error);
+            //model.addAttribute("comments", image.getComments());
+            return "images/image";
+        }
     }
